backend/requirements.txt
fastapi
uvicorn
pinecone-client
sentence-transformers
requests
python-dotenv
pytest

